more...
this is actually a serious problem.
Our repeated code in generation makes it so that the attention matrices
generated are inconsistent. For example, for the new function you made
to extract tokenwise attentions you were preprocessing it on the spot,
which is not supposed to, because we already process it on
graph_feature. this leads to double preprocessing and incorrect results
+ harder to track data flow.